Tested with MANY items and made a few tweaks
Mainly database corrections and fixes but a few code changes.

Affixes are now limited to the item level.

The next work will be to remove implicit mods (carefully) from some
items once parsed as these incorrectly report an implicit mod is left
over (not all affixes were assigned error).  Some are tricky, e.g. a
weapon with +accuracy as an affix and an implicit.