Added Location
Location stores the place an item is stored for the stash viewer.

Also vastly improved the uniqueness test for stashed items so it makes a
lot more sense and tries to get the best information possible, e.g.
updates the item level and location if they are better than the ones
stored already.

This means you can use Ctrl-C to update your stashed items' level where
you feel the need.  If you have the option set to not allow duplicates
in the stash (in the settings).